feat: call increaseRewardsAdjustment of dashboard when consolidation reqeust triggers, fix tests, add new, rename contract and event
