+++ lib/protocol/types.ts
@@ -20,7 +20,6 @@ import {
   Lido,
   LidoExecutionLayerRewardsVault,
   LidoLocator,
-  MaxEffectiveBalanceIncreaser,
   NodeOperatorsRegistry,
   OperatorGrid,
   OracleDaemonConfig,
@@ -28,6 +27,7 @@ import {
   PredepositGuarantee,
   StakingRouter,
   UpgradeableBeacon,
+  ValidatorConsolidationRequests,
   ValidatorsExitBusOracle,
   VaultFactory,
   VaultHub,
@@ -71,7 +71,7 @@ export type ProtocolNetworkItems = {
   vaultHub: string;
   predepositGuarantee: string;
   operatorGrid: string;
-  maxEffectiveBalanceIncreaser: string;
+  validatorConsolidationRequests: string;
   lazyOracle: string;
 };
 
@@ -101,7 +101,7 @@ export interface ContractTypes {
   OperatorGrid: OperatorGrid;
   IStakingModule: IStakingModule;
   ICSModule: ICSModule;
-  MaxEffectiveBalanceIncreaser: MaxEffectiveBalanceIncreaser;
+  ValidatorConsolidationRequests: ValidatorConsolidationRequests;
   LazyOracle: LazyOracle;
 }
 
@@ -157,7 +157,7 @@ export type VaultsContracts = {
   vaultHub: LoadedContract<VaultHub>;
   predepositGuarantee: LoadedContract<PredepositGuarantee>;
   operatorGrid: LoadedContract<OperatorGrid>;
-  maxEffectiveBalanceIncreaser: LoadedContract<MaxEffectiveBalanceIncreaser>;
+  validatorConsolidationRequests: LoadedContract<ValidatorConsolidationRequests>;
   lazyOracle: LoadedContract<LazyOracle>;
 };
