+++ scripts/scratch/steps.json
@@ -17,7 +17,6 @@
     "scratch/steps/0120-post-locator-initializers",
     "scratch/steps/0130-grant-roles",
     "scratch/steps/0140-plug-staking-modules",
-    "scratch/steps/0145-deploy-vaults",
     "scratch/steps/0150-transfer-roles",
     "scratch/steps/0160-deploy-max-effective-balance-increaser"
   ]
