+++ test/0.8.25/vaults/contracts/DelegateCaller.sol
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: UNLICENSED
 // for testing purposes only
 
-pragma solidity 0.8.9;
+pragma solidity 0.8.25;
 
 contract DelegateCaller {
     function callDelegate(address target, bytes memory data) external payable returns (bytes memory) {
