+++ test/0.8.25/vaults/contracts/EIP7251MaxEffectiveBalanceRequest__Mock.sol
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: UNLICENSED
 // for testing purposes only
 
-pragma solidity 0.8.9;
+pragma solidity 0.8.25;
 
 /**
  * @notice This is a mock of EIP-7251's pre-deploy contract.
